base/tg-session/requirements.txt
telethon==1.42.0
qrcode==8.2
python-dotenv==1.2.2